# src/views/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MailScheduler API project",
        "version": "v1"
    },
    "paths": {
        "/api/Emails": {
            "get": {
                "tags": [
                    "Emails"
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Emails"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Emails"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EmailModel"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/Emails/{id}": {
            "delete": {
                "tags": [
                    "Emails"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "EmailModel": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "object",
                        "properties": {
                            "id": {
                                "type": "string"
                            },
                            "to": {
                                "type": "string"
                            },
                            "cc": {
                                "type": "string"
                            },
                            "ccn": {
                                "type": "string"
                            },
                            "subject": {
                                "type": "string"
                            },
                            "body": {
                                "type": "string"
                            },
                            "job": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "cron": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "InvalidResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
}